requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
loguru==0.7.2
motor==3.3.2
packaging==25.0
Pillow==10.0.1
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyJWT==2.8.0
pymongo==4.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
